staging/pkg/mod/golang.org/x/net@v0.0.0-20210520170846-37e1c6afe023/html/testdata/webkit/tests23.dat

#data
<p><font size=4><font color=red><font size=4><font size=4><font size=4><font size=4><font size=4><font color=red><p>X
#errors
(1,3): expected-doctype-but-got-start-tag
(1,116): unexpected-end-tag
(1,117): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <font>
|         size="4"
|         <font>
|           color="red"
|           <font>
|             size="4"
|             <font>
|               size="4"
|               <font>
|                 size="4"
|                 <font>
|                   size="4"
|                   <font>
|                     size="4"
|                     <font>
|                       color="red"
|     <p>
|       <font>
|         color="red"
|         <font>
|           size="4"
|           <font>
|             size="4"
|             <font>
|               size="4"
|               <font>
|                 color="red"
|                 "X"

#data
<p><font size=4><font size=4><font size=4><font size=4><p>X
#errors
(1,3): expected-doctype-but-got-start-tag
(1,58): unexpected-end-tag
(1,59): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <font>
|         size="4"
|         <font>
|           size="4"
|           <font>
|             size="4"
|             <font>
|               size="4"
|     <p>
|       <font>
|         size="4"
|         <font>
|           size="4"
|           <font>
|             size="4"
|             "X"

#data
<p><font size=4><font size=4><font size=4><font size="5"><font size=4><p>X
#errors
(1,3): expected-doctype-but-got-start-tag
(1,73): unexpected-end-tag
(1,74): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <font>
|         size="4"
|         <font>
|           size="4"
|           <font>
|             size="4"
|             <font>
|               size="5"
|               <font>
|                 size="4"
|     <p>
|       <font>
|         size="4"
|         <font>
|           size="4"
|           <font>
|             size="5"
|             <font>
|               size="4"
|               "X"

#data
<p><font size=4 id=a><font size=4 id=b><font size=4><font size=4><p>X
#errors
(1,3): expected-doctype-but-got-start-tag
(1,68): unexpected-end-tag
(1,69): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <font>
|         id="a"
|         size="4"
|         <font>
|           id="b"
|           size="4"
|           <font>
|             size="4"
|             <font>
|               size="4"
|     <p>
|       <font>
|         id="a"
|         size="4"
|         <font>
|           id="b"
|           size="4"
|           <font>
|             size="4"
|             <font>
|               size="4"
|               "X"

#data
<p><b id=a><b id=a><b id=a><b><object><b id=a><b id=a>X</object><p>Y
#errors
(1,3): expected-doctype-but-got-start-tag
(1,64): end-tag-too-early
(1,67): unexpected-end-tag
(1,68): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <b>
|         id="a"
|         <b>
|           id="a"
|           <b>
|             id="a"
|             <b>
|               <object>
|                 <b>
|                   id="a"
|                   <b>
|                     id="a"
|                     "X"
|     <p>
|       <b>
|         id="a"
|         <b>
|           id="a"
|           <b>
|             id="a"
|             <b>
|               "Y"
